Squashed commit of the following:
commit 73cce3df5b2dd4301fbea3670ee8c8ea8cd9a98c
Author: Mike Fahey <michaelpfahey@gmail.com>
Date:   Mon Nov 10 14:50:03 2014 -0800

Improved string representation of array values in jmx connector

Arrays were being rendered using the default toString representation.
So, an Array would be rendered as something like [Ljava.lang.String;@4da91ced.
Arrays are now rendered as a string representation of the contents of the array.
The string representation consists of a list of the array's elements,
enclosed in square brackets ("[]").  Adjacent elements are separated by ", ".

Fixes #1632